requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain_community==0.0.20
tiktoken==0.5.2
langchainhub==0.1.14
chromadb==0.4.22
python-dotenv==0.19.0